backend-example/requirements.txt
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
Deprecated==1.2.15
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-core==2.4.1
google-cloud-logging==3.11.3
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
idna==3.10
importlib_metadata==8.5.0
opentelemetry-api==1.29.0
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
requests==2.32.3
rsa==4.9
urllib3==2.3.0
wrapt==1.17.1
zipp==3.21.0
